+++ src/js/index.js
@@ -1,6 +1,7 @@
 // All helper functions will be imported here, so that they can all be exported within one object.
-import { getBreakpoints } from './modules/breakpointHelper';
+import { getBreakpoints, currentScreenWidth } from './modules/breakpointHelper';
 
 export {
-    getBreakpoints
+    getBreakpoints,
+    currentScreenWidth
 };